fix bugs and add tests, test dir moved
minor changes in endpoints (currency and user), handlers.py

bugs fixed in database.py, security.py

tests added, tests dir moved from app to the root, added fixture to conftest.py

changed types of attributes in currency schemas

+++ app/api/endpoints/currency.py
@@ -25,7 +25,7 @@ currency_router = APIRouter(
 async def currency_exchange(
     currency_1: Annotated[str, Query(..., alias="from")],
     currency_2: Annotated[str, Query(..., alias="to")],
-    amount: int,
+    amount: float,
 ) -> CurrencyResponse:
     params = locals()
     return ext_api_get_exchange(CurrencyRequest(**params))
